Move ValidationResult into validation module
