+++ core/src/action.rs
@@ -5,7 +5,8 @@ use crate::{
         direct_message::DirectMessage, entry_with_header::EntryWithHeader, state::NetworkState,
     },
     nucleus::{
-        state::{NucleusState, ValidationResult},
+        state::NucleusState,
+        validation::ValidationResult,
         ExecuteZomeFnResponse, ZomeFnCall,
     },
     scheduled_jobs::pending_validations::PendingValidation,
