+++ core/src/nucleus/state.rs
@@ -1,4 +1,7 @@
-use crate::{nucleus::ZomeFnCall, scheduled_jobs::pending_validations::PendingValidation};
+use crate::{
+    nucleus::{validation::ValidationResult, ZomeFnCall},
+    scheduled_jobs::pending_validations::PendingValidation,
+};
 use holochain_core_types::{
     cas::content::Address, dna::Dna, error::HolochainError, json::JsonString,
     validation::ValidationPackage,
@@ -17,30 +20,6 @@ pub enum NucleusStatus {
 impl Default for NucleusStatus {
     fn default() -> Self {
         NucleusStatus::New
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum ValidationError {
-    Fail(String),
-    UnresolvedDependencies(Vec<Address>),
-    NotImplemented,
-    Error(String),
-}
-pub type ValidationResult = Result<(), ValidationError>;
-
-impl From<ValidationError> for HolochainError {
-    fn from(ve: ValidationError) -> Self {
-        match ve {
-            ValidationError::Fail(reason) => HolochainError::ValidationFailed(reason),
-            ValidationError::UnresolvedDependencies(_) => {
-                HolochainError::ValidationFailed("Missing dependencies".to_string())
-            }
-            ValidationError::NotImplemented => {
-                HolochainError::NotImplemented("Validation not implemented".to_string())
-            }
-            ValidationError::Error(e) => HolochainError::ErrorGeneric(e),
-        }
     }
 }
