+++ core/src/nucleus/validation/header_address.rs
@@ -1,4 +1,4 @@
-use crate::nucleus::state::{ValidationError, ValidationResult};
+use crate::nucleus::validation::{ValidationError, ValidationResult};
 use boolinator::Boolinator;
 use holochain_core_types::{
     cas::content::AddressableContent, chain_header::ChainHeader, entry::Entry,
